Sampling and event data
Sampling and event data provided originally by HCMR as a google sheet, which a year later was deleted and we were told not to use these data because not all stations should be published. However, the data in this spreadsheet were not made available elsewhere, so we commit here a copy that we had made of that spreadsheet when it was first provided to us.